Move interchain_send_tx fn to test-framework crate
