+++ tools/integration-test/src/tests/ica.rs
@@ -23,8 +23,7 @@ use ibc_test_framework::prelude::*;
 use ibc_test_framework::relayer::channel::{
     assert_eventually_channel_established, query_channel_end,
 };
-
-use crate::tests::utils::interchain_send_tx;
+use ibc_test_framework::util::interchain_security::interchain_send_tx;
 
 #[test]
 fn test_ica_filter_default() -> Result<(), Error> {
