+++ tools/integration-test/src/tests/interchain_security/mod.rs
@@ -1,6 +1,6 @@
-pub mod ica_ordered_channel;
 #[cfg(any(doc, feature = "ica"))]
 pub mod ica_transfer;
+pub mod ica_ordered_channel;
 #[cfg(any(doc, feature = "ics31"))]
 pub mod icq;
 pub mod simple_transfer;
