+++ tools/integration-test/src/tests/mod.rs
@@ -25,8 +25,6 @@ pub mod tendermint;
 pub mod ternary_transfer;
 pub mod transfer;
 
-mod utils;
-
 #[cfg(any(doc, feature = "ics29-fee"))]
 pub mod fee;
